# api-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API (Go/Fiber). Документация и Try it out: GET /docs/index.html. Для цепочки slug в path используй %2F (например elektronika%2Ftelefony).\n**Socket.IO:** тот же хост и CORS что у REST; путь `/socket.io`; авторизация cookie `session_id`. Неймспейсы `/chat` и `/support`.\nКлиент `socket.io`: надёжнее **v2.x / v3.x**; движок Go (`googollee/go-socket.io` ~v1.7) часто **не коннектится** с **socket.io-client v4** из‑за протокола handshake.\n**/chat** — исходящие: `joinChat` {chatId}, `sendMessage` {chatId, content}, `markAsRead` {chatId}. Входящие: `newMessage`, `messagesRead`, `newChatMessage`.\n**/support** — исходящие: `joinTicket`, `leaveTicket`, `sendSupportMessage` {ticketId, message:{text}}, `supportTyping` {ticketId, isTyping}, `updateTicketStatus` {ticketId, status} (модератор). Входящие: `newSupportMessage`, `moderatorResponse`, `supportUserTyping`, `ticketStatusUpdated`; по неймспейсу `newSupportTicket`; в комнате тикета `ticketAssigned`.",
        "title": "Med Vito REST API (Go)",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/support/tickets": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "support"
                ],
                "summary": "Создать тикет поддержки",
                "parameters": [
                    {
                        "description": "Тело",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.swaggerCreateSupportTicket"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/support/tickets/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "support"
                ],
                "summary": "Тикет с сообщениями",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "support"
                ],
                "summary": "Обновить тикет (модератор/admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Поля",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.swaggerUpdateSupportTicket"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/support/tickets/{id}/assign": {
            "put": {
                "tags": [
                    "support"
                ],
                "summary": "Назначить тикет на себя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/support/tickets/{id}/messages": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "support"
                ],
                "summary": "Сообщение в тикет",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Текст",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.swaggerSupportMessage"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "main.swaggerCreateSupportTicket": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "priority": {
                    "type": "string",
                    "example": "MEDIUM"
                },
                "subject": {
                    "type": "string"
                },
                "theme": {
                    "type": "string",
                    "example": "TECHNICAL_ISSUE"
                }
            }
        },
...
        "main.swaggerSupportMessage": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
...
        "main.swaggerUpdateSupportTicket": {
            "type": "object",
            "properties": {
                "priority": {
                    "type": "string",
                    "example": "HIGH"
                },
                "status": {
                    "type": "string",
                    "example": "IN_PROGRESS"
                }
            }
        },
...
    },
...
}